use static cast

+++ oneflow/core/framework/random_generator_impl.cu
@@ -27,8 +27,8 @@ __global__ void InitCurandStatesKernel(uint64_t seed, curandState* states, uint6
   size_t local_seed = (static_cast<size_t>(seed) + 0x9e3779b9U + (static_cast<size_t>(id) << 6U)
                        + (static_cast<size_t>(id) >> 2U));
   curand_init(local_seed, 0, 0, &states[id]);
-  *dev_counter = 0; 
-  *dev_offset = 0; 
+  *dev_counter = static_cast<uint64_t>(0); 
+  *dev_offset = static_cast<uint64_t>(0); 
 }
 
 }  // namespace
